refactor(r2dreamer): split agent into world_model/behavior/representation packages
agent.py was a 786-LOC file whose 200-LOC _loss_fn() simultaneously computed
world-model KL, Barlow-Twins, reward & continue heads, imagination rollout,
actor-critic gradients, and replay-based value learning. Every gradient flowed
through one jax.grad(), so isolating any one loss for an ablation required
forking the whole function.

The loss math now lives in three subpackages:

  world_model/    rssm, encoders, heads, KL+reward+continue loss
  behavior/       return_ema, imagination, actor+critic loss
  representation/ Barlow projector+loss, replay value loss

agent.py is now a composition root: a shared forward pass (encoder + RSSM
observe + prior + feat) feeds the three sub-loss functions, which are summed
under a single jax.grad. Toggling a loss is now a one-line change at the
composition root rather than surgery inside _loss_fn.

networks.py is a re-export shim so existing imports keep working; _kl_loss,
_lambda_return, and _imagine remain importable from agent for the parity tests.

Verified: 67/67 modules/r2dreamer/tests pass, including the 14
test_cross_framework checks that pin numeric parity against PyTorch.
