+++ modules/r2dreamer/representation/__init__.py
@@ -1,0 +1,7 @@
+"""Representation subpackage: Barlow Twins + replay-based value learning."""
+
+from .barlow import Projector, barlow_loss
+from .repvalue import repval_loss
+from .loss import representation_loss
+
+__all__ = ["Projector", "barlow_loss", "repval_loss", "representation_loss"]
