+++ modules/r2dreamer/world_model/__init__.py
@@ -1,0 +1,13 @@
+"""World-model subpackage: encoders, RSSM, head primitives, and world-model loss."""
+
+from .rssm import RMSNorm, BlockLinear, Deter, R2RSSM
+from .encoders import R2Encoder, VGGTEncoder, VGGTAggregatorMLPEncoder
+from .heads import R2MLP, R2TwoHotDist, onehot_mode_st
+from .loss import world_model_loss, kl_loss
+
+__all__ = [
+    "RMSNorm", "BlockLinear", "Deter", "R2RSSM",
+    "R2Encoder", "VGGTEncoder", "VGGTAggregatorMLPEncoder",
+    "R2MLP", "R2TwoHotDist", "onehot_mode_st",
+    "world_model_loss", "kl_loss",
+]
